Add float number

+++ app/src/main/java/com/example/logointerpreterbeta/Turtle.java
@@ -6,7 +6,7 @@ import android.graphics.Color;
 public class Turtle {
     public static float Xposition = 500;
     public static float Yposition = 500;
-    public static int direction = 0;
+    public static float direction = 0;
     public static boolean isDown = true;
     public static int penColor = Color.WHITE;
     public static int penSize = 5;
